restructure func to parse jwt
