+++ helper/jwts/get_token_jwt.go
@@ -1,4 +1,4 @@
-package middlewares
+package jwts
 
 import (
 	"strings"
