+++ internal/middlewares/jwt_claims.go
@@ -1,7 +1,6 @@
 package middlewares
 
 import (
-	"log"
 	"os"
 
 	"github.com/golang-jwt/jwt/v4"
@@ -18,68 +17,4 @@ type JWTClaim struct {
 	Username string
 	Role     string
 	jwt.RegisteredClaims
-}
-
-func GetID(token string) (string, error) {
-	claims := &JWTClaim{}
-
-	// parsing token jwt
-	_, err := jwt.ParseWithClaims(token, claims, func(t *jwt.Token) (interface{}, error) {
-		return []byte(JWT_SECRET_KEY), nil
-	})
-
-	if err != nil {
-		log.Println(err)
-		return "", err
-	}
-
-	return claims.ID, nil
-}
-
-func GetUsername(token string) (string, error) {
-	claims := &JWTClaim{}
-
-	// parsing token jwt
-	_, err := jwt.ParseWithClaims(token, claims, func(t *jwt.Token) (interface{}, error) {
-		return []byte(JWT_SECRET_KEY), nil
-	})
-
-	if err != nil {
-		log.Println(err)
-		return "", err
-	}
-
-	return claims.Username, nil
-}
-
-func GetRole(token string) (string, error) {
-	claims := &JWTClaim{}
-
-	// parsing token jwt
-	_, err := jwt.ParseWithClaims(token, claims, func(t *jwt.Token) (interface{}, error) {
-		return []byte(JWT_SECRET_KEY), nil
-	})
-
-	if err != nil {
-		log.Println(err)
-		return "", err
-	}
-
-	return claims.Role, nil
-}
-
-func GetClaims(token string) (*JWTClaim, error) {
-	claims := &JWTClaim{}
-	
-	// parsing token jwt
-	_, err := jwt.ParseWithClaims(token, claims, func(t *jwt.Token) (interface{}, error) {
-		return []byte(JWT_SECRET_KEY), nil
-	})
-
-	if err != nil {
-		log.Println(err)
-		return nil, err
-	}
-
-	return claims, nil
 }
